Add empty state for auth hooks page
