+++ apps/studio/pages/project/[ref]/auth/hooks.tsx
@@ -3,6 +3,7 @@ import { PermissionAction } from '@supabase/shared-types/out/constants'
 import { HooksListing } from 'components/interfaces/Auth/Hooks/HooksListing'
 import { AuthLayout } from 'components/layouts'
 import { ScaffoldContainer } from 'components/layouts/Scaffold'
+import { FormsContainer } from 'components/ui/Forms'
 import NoPermission from 'components/ui/NoPermission'
 import { useCheckPermissions, usePermissionsLoaded } from 'hooks'
 import type { NextPageWithLayout } from 'types'
@@ -15,9 +16,9 @@ const Hooks: NextPageWithLayout = () => {
     return <NoPermission isFullPage resourceText="access your project's auth hooks" />
   } else {
     return (
-      <ScaffoldContainer className="pt-12 pb-20">
+      <FormsContainer>
         <HooksListing />
-      </ScaffoldContainer>
+      </FormsContainer>
     )
   }
 }
